refactor: Update ThreadList import path and remove obsolete file
